cleanup inbound addresses queries and unused estimate hook
