+++ src/pages/ThorChainLP/components/RemoveLiquidity/RemoveLiquidityInput.tsx
@@ -24,14 +24,11 @@ import type { AccountId } from '@shapeshiftoss/caip'
 import { thorchainAssetId, thorchainChainId, toAccountId } from '@shapeshiftoss/caip'
 import { SwapperName } from '@shapeshiftoss/swapper'
 import type { Asset, MarketData } from '@shapeshiftoss/types'
-import { useQuery } from '@tanstack/react-query'
 import React, { useCallback, useEffect, useMemo, useState } from 'react'
 import { BiSolidBoltCircle } from 'react-icons/bi'
 import { FaPlus } from 'react-icons/fa6'
 import { useTranslate } from 'react-polyglot'
-import { reactQueries } from 'react-queries'
 import { useIsTradingActive } from 'react-queries/hooks/useIsTradingActive'
-import { selectInboundAddressData } from 'react-queries/selectors'
 import { useHistory } from 'react-router'
 import { Amount } from 'components/Amount/Amount'
 import { AssetInput } from 'components/DeFi/components/AssetInput'
@@ -50,12 +47,8 @@ import { getMixPanel } from 'lib/mixpanel/mixPanelSingleton'
 import { MixPanelEvent } from 'lib/mixpanel/types'
 import { THORCHAIN_OUTBOUND_FEE_RUNE_THOR_UNIT } from 'lib/swapper/swappers/ThorchainSwapper/constants'
 import { assertUnreachable } from 'lib/utils'
-import {
-  fromThorBaseUnit,
-  getThorchainFromAddress,
-  getThorchainTransactionType,
-} from 'lib/utils/thorchain'
-import { THOR_PRECISION, THORCHAIN_POOL_MODULE_ADDRESS } from 'lib/utils/thorchain/constants'
+import { fromThorBaseUnit, getThorchainFromAddress } from 'lib/utils/thorchain'
+import { THOR_PRECISION } from 'lib/utils/thorchain/constants'
 import { useSendThorTx } from 'lib/utils/thorchain/hooks/useSendThorTx'
 import { estimateRemoveThorchainLiquidityPosition } from 'lib/utils/thorchain/lp'
 import type { LpConfirmedWithdrawalQuote, UserLpDataPosition } from 'lib/utils/thorchain/lp/types'
@@ -181,19 +174,6 @@ export const RemoveLiquidityInput: React.FC<RemoveLiquidityInputProps> = ({
   const runeBalanceCryptoBaseUnit = useAppSelector(state =>
     selectPortfolioCryptoBalanceBaseUnitByFilter(state, runeBalanceFilter),
   )
-
-  const { data: inboundAddressesData } = useQuery({
-    ...reactQueries.thornode.inboundAddresses(),
-    select: data => selectInboundAddressData(data, assetId),
-    // @lukemorales/query-key-factory only returns queryFn and queryKey - all others will be ignored in the returned object
-    // Go stale instantly
-    staleTime: 0,
-    // Never store queries in cache since we always want fresh data
-    gcTime: 0,
-    refetchOnWindowFocus: true,
-    refetchOnMount: true,
-    refetchInterval: 60_000,
-  })
 
   const { isTradingActive, isLoading: isTradingActiveLoading } = useIsTradingActive({
     assetId: poolAsset?.assetId,
@@ -409,6 +389,7 @@ export const RemoveLiquidityInput: React.FC<RemoveLiquidityInputProps> = ({
     isEstimatedFeesDataLoading: isEstimatedPoolAssetFeesDataLoading,
     isEstimatedFeesDataError: isEstimatedPoolAssetFeesDataError,
     dustAmountCryptoBaseUnit: poolAssetFeeAssetDustAmountCryptoBaseUnit,
+    outboundFeeCryptoBaseUnit,
   } = useSendThorTx({
     assetId: poolAsset?.assetId,
     accountId,
@@ -420,9 +401,10 @@ export const RemoveLiquidityInput: React.FC<RemoveLiquidityInputProps> = ({
   })
 
   const poolAssetProtocolFeeCryptoPrecision = useMemo(() => {
+    if (!poolAssetFeeAsset || !outboundFeeCryptoBaseUnit) return bn(0)
     if (bnOrZero(actualAssetWithdrawAmountCryptoPrecision).eq(0)) return bn(0)
-    return fromThorBaseUnit(inboundAddressesData?.outbound_fee ?? '0')
-  }, [inboundAddressesData?.outbound_fee, actualAssetWithdrawAmountCryptoPrecision])
+    return bnOrZero(fromBaseUnit(outboundFeeCryptoBaseUnit, poolAssetFeeAsset.precision))
+  }, [outboundFeeCryptoBaseUnit, actualAssetWithdrawAmountCryptoPrecision, poolAssetFeeAsset])
 
   const poolAssetProtocolFeeFiatUserCurrency = useMemo(() => {
     return poolAssetProtocolFeeCryptoPrecision.times(poolAssetFeeAssetMarketData.price)
@@ -500,28 +482,6 @@ export const RemoveLiquidityInput: React.FC<RemoveLiquidityInputProps> = ({
       </Flex>
     )
   }, [])
-
-  const poolAssetInboundAddress = useMemo(() => {
-    if (!poolAsset) return
-
-    const transactionType = getThorchainTransactionType(poolAsset.chainId)
-
-    switch (transactionType) {
-      case 'MsgDeposit':
-        return THORCHAIN_POOL_MODULE_ADDRESS
-
-      case 'EvmCustomTx':
-        // TODO: this should really be inboundAddressData?.router, but useQuoteEstimatedFeesQuery doesn't yet handle contract calls
-        // for the purpose of naively assuming a send, using the inbound address instead of the router is fine
-        return inboundAddressesData?.address
-
-      case 'Send':
-        return inboundAddressesData?.address
-
-      default:
-        assertUnreachable(transactionType as never)
-    }
-  }, [poolAsset, inboundAddressesData?.address])
 
   const renderHeader = useMemo(() => {
     if (headerComponent) return headerComponent
@@ -636,7 +596,6 @@ export const RemoveLiquidityInput: React.FC<RemoveLiquidityInputProps> = ({
     if (!actualRuneWithdrawAmountCryptoPrecision) return
     if (!actualRuneWithdrawAmountFiatUserCurrency) return
     if (!shareOfPoolDecimalPercent) return
-    if (!poolAssetInboundAddress) return
 
     setConfirmedQuote({
       assetWithdrawAmountCryptoPrecision: actualAssetWithdrawAmountCryptoPrecision,
@@ -646,7 +605,6 @@ export const RemoveLiquidityInput: React.FC<RemoveLiquidityInputProps> = ({
       shareOfPoolDecimalPercent,
       slippageFiatUserCurrency,
       opportunityId,
-      quoteInboundAddress: poolAssetInboundAddress,
       runeGasFeeFiatUserCurrency: runeGasFeeFiatUserCurrency.toFixed(2),
       poolAssetGasFeeFiatUserCurrency: poolAssetGasFeeFiatUserCurrency.toFixed(2),
       totalGasFeeFiatUserCurrency: totalGasFeeFiatUserCurrency.toFixed(2),
@@ -666,7 +624,6 @@ export const RemoveLiquidityInput: React.FC<RemoveLiquidityInputProps> = ({
     percentageSelection,
     poolAsset,
     poolAssetGasFeeFiatUserCurrency,
-    poolAssetInboundAddress,
     position,
     runeAccountId,
     runeGasFeeFiatUserCurrency,
